# CUDA: parallelize graph evaluation across multiple streams