@sorasoras 2024-01-01
I tried to compile with your branch with rocm on windows but it does not seems to support it.