@sorasoras 2024-01-01
> What specifically is failing? Compilation? If yes, please provide a log.

o, nevermind, just wrong conf.
But with your pr, I only get like half utilization of the main branch with half speed. With main branch, my  7900XTX can hit 420~W, but with this pr, i could only ever reach 200w.